Show a list of companies.
Initially when the user hasn't
started a search, I'll show
all listings. I want the user
to be aware that this page is
showing them a list of companies.

Once they start typing, I do a
debounced search. In a production
environment, we probably wouldn't
want to be calling the API on every
key press b/c it would slow down
the page.

The user can also click the search
button to do a search as well.