Implemented paginated design
Now users can switch b/w pages.
There's a first, prev, next,
and last button in a nav
section at the bottom of the page.

When switching to the next page,
the app will auto-scroll back
to the top of the page.

I also remove the Search button,
b/c I want the search field to
be aligned in the center of the page.
The button was cosmetic, and didn't
do anything.