Auto-fill space if no image
If there's no image, the
remaining part of the card should
fill the empty space.

I also fix the width of the image,
and the height of the card, so
they look a little more uniform
in layout and size. The image
shouldn't take up more than a
quarter of the card. IMO, I'm
just entertaining myself now.