Reset page # when query changes
I found an annoying issue that
ultimately helped me learn a bit
more about React.

When you update your query, I want
the page # to reset. For example,
if you're on page 2 with the query
"ba", and then type "bar", the list
should reset back to page 1 with the
new search results.

This wasn't happening though, b/c
the component's position in the
render tree didn't change. If
the page # was 2, and you updated
the query state of the parent,
the list was still in the same
position and so page # would stay the same.
https://react.dev/learn/preserving-and-resetting-state

If I want to reset the pageNo state in
the child, I have to give it a key on the
query. This will tell React that there's
a new component that it needs to render
in the tree.