Add error handling
Before this commit,
I was having my API
functions throw errors
if the API didn't return
a successful result.

This wouldn't work in a
production scenario, b/c
we don't want to crash the page.
Instead, I added error handling.

In the case of getFavoritesCount()
and getListings(), it's sufficient
to return 0 or nothing and tell
the user "No results found".

In the case of changing the star
status though, I needed to have
it return a boolean to indicate
success or failure, and then
only make the corresponding
state changes across the app
if the call was successful.

If the call to change the starred
status is not successful, then
I show the user an error message
and do not change any state. A
window alert is not great, but
for the interest of time that's
what I'll use.