Reset pageNo in separate useEffect
The downside of using a key to remount
the entire list component is that it causes
layout shifts when typing in input and
changing your search.

That is not ideal from a UX perspective:
the user shouldn't be jerked around on
the page. An alternative approach is
to reset pageNo on query change with a
separate useEffect.